here are the lastest changes still working on it.
@@ -44,13 +44,13 @@ Package management
- setting up my self written feature file and load it just was not sure
the proper syntax, or which is the command that I need for this. The
whole not needing the absolute address does not seem to work for
anything I try so that was a bit of bumble fucking around to get it to
work

@@ -122,7 +122,10 @@ startup-settings
- first is I found a mouse avoidance mode which I have turned on and I
like it so far have yet to have any problem with it and it works and I
hardly notice that is a good setting or package or app that does that so
I am going to keep it.
- the next thing I thought was a useless setting and removed it and then
I realized that no it was doing something so I brought it back.

@@ -160,7 +163,8 @@ dashboard
- here is were I install and use the feature file that I made. It'll
open a workspace for the config named and able to quit with just "q"

@@ -262,52 +266,6 @@ org-settings
- this was left behind after it was moved to it's own heading ..purge

@@ -609,6 +567,32 @@ evil-surround
- here I'm working on a method for moving around code, what a cunvaluted
space, you have all things converging here. You have multiple ways of
doing it and some don't seem to work while others do add there is so
many things involved evil-surround, smartparens, paraedit, evil-motions,
emacs-defaults, the smartparens key interferes with arch and my move
workstations. what I have read doom has deactivated the
smartparens-motion, the reason he gave made sense it was that
smartparens-motion was expensive and there was already a system implace
to this and most were already familiar with. rather then learning yet
another set of motion keys. So that were I end up enabling this which is
evil-embrace intergrated into evil-surround. Still have problems with
this but need to use a trial and error test to figure out what works.

@@ -967,9 +951,64 @@ consult
- removed this earlier, but now want it back after, if nothing else then
as a Reference, it's commited out for now will look at another time see
what I want to keep.

@@ -1469,10 +1508,14 @@ added-pkgs
- here I added a package "wrap search" willing to try it out it small
enough to just manual load and it has potential to be used rather that
much bigger and more complicated packages. will see

@@ -1494,9 +1537,10 @@ org-rich-yank
- this nested prefix and all nested prefixes throw warnings so ...purge

@@ -2762,44 +2817,26 @@ denote
- still warming up to denote yet, now there is a new version and a wack
of updates it hard to keep up but here is the lastest rendition, there
are a lot of new features, like dynamic-blocks for backlinking lists.
many more that was just one that jumped out at me, still getting use to
it now there is more.

@@ -2812,14 +2849,15 @@ Denote
spent small amount of time looking at this new app. seems like it could
be used. but was not working properly. It seem to have to do with evil
and the keys it grabs and the different function that it out puts so I
tried a little trick I have seen used in my travels and that was just to
remap that key for that mode. seem to work so happy with that and there
seems to be no side effect so thats.

@@ -2938,13 +2976,13 @@ typing-exercise
- fine tuning