work in progress update
- this is to update the work done so far

@@ -64,7 +64,7 @@ Fonts
- after an update. flycheck told me to change this so there it is

@@ -91,7 +91,10 @@ startup-settings
- this caused some problems when entered wrong in bashrc so just going
to try and leave ENVIRONMENTS alone

@@ -397,7 +400,7 @@ org-capture
- try correcting this

@@ -438,16 +441,17 @@ org-journal
- this is the start of more changes for journal
- added a link to Doom-Modules-Configuration
- no need to declare the package since Doom is doing that already
- don't believe I need to declare modern-mode
@@ -466,6 +470,14 @@ org-journal
- here I started grabbing settings that were just in customize-menu
- note did not realize that removing the require also removed the
previous hooks not sure if this was a one off cause I found it were it
should be in the custom.el-file. But this is the better way to do it.
* don't use customize-menu. just put it in the configuration.

@@ -647,18 +659,20 @@ key-chords
- Doom defer and ! macro, removed keys that now I tend to use the
defaults so have not use these much ... purge

@@ -668,11 +682,11 @@ scroll-margin
- wanted to see what this was doing and I not sure still and it's been
off for a while now so I am not seeing what purpose this is doing.
will leave as so for now but later might as well purge it to.

@@ -1269,7 +1283,7 @@ fly-spell
- not sure what this is doing with flyspell but will purge in future

@@ -1568,7 +1582,8 @@ my-keybindings
- binding for wrap-search not sure how long I will keep this, seems
redundant since my original settings work pretty bad ass as is this just
adds another option that really isn't necessary.

@@ -1577,6 +1592,18 @@ my-keybindings
- Don't even get me started on this .... lets just say this is
work-in-progress and this is mostly trial and error so fuck off with the
whole parentheses fucking thing! OK! Great! Fine! It's fine I said. Now
fuck off about it already. Never mind, just move on.

@@ -1841,27 +1866,6 @@ Continues comments if executed from a commented line. Consults
- not needed ... purge

@@ -2315,6 +2324,8 @@ Elfeed
- little thing to open summary everytime instead doing manually

@@ -3027,33 +3051,3 @@ view-text-as-info
this was interesting and all thou I did get it to work first it was what
I expected and when it did this it left the files in the folders and
threw an error so I archived it if need in future