aesthetics and the adding of a elfeed-org and yequake
feel the need to clean up version control, I find myself using it more
and would like it to be less cruft. This had the bonus effect of making
my commits smaller. This was something I wanted any way so it is a win
win. Most should not need an explanation cause they are for aesthetics.
there is two things of any importance. adding the elfeed-org section
which is no big difference since nothing else changed other than it has
it's own header now. the other thing is yequake-frames.

@@ -124,8 +124,6 @@ startup-settings
- this setting did not do any thing good. infact it caused some problems
so after looking in the documentation I concluded that I did not need
this. Purge!

@@ -401,24 +399,28 @@ org-capture-templates
- This is my attempt to take back control of my capture-templates, still
need to go to the customize-menu to lock it in but I'm working on that.

@@ -2388,9 +2391,30 @@ Elfeed
- this is the new place for elfeed-org

@@ -2923,7 +2948,7 @@ denote
- this setting should help denote files to appear the same as the rest
without the timestamp in the front of it.

@@ -3092,3 +3117,43 @@ ediff
- And here it is. This is yequake-frames. It will replace conkey for
quick popup buffer that I kept my own personal key reference. This is a
big improvement over the old method. The only draw back is it has been
deficit for a while and I have since found other way of obtaining the
same info. So time will tell if I start using it more or not. ether way
it is here to stay, for what ever reason I want to use it for.