need to clean up the diff so "update"
@@ -1,9 +1,10 @@
- this was what I use to call a preable then changed to properties. But
at the end of the day it was a properties-drawer and everything was fine
till someone pointed out something in a forum it was going on about the
proper syntax for properties and some thing were not suppose to be in
the properties-drawer. This made me rethink things and redo this all
bringing out things that were more keywords like #+startup: and so on.
After doing this first I seen no difference in how it is acting and the
only thing that changed is now I have all this clutter at the top of my
file I didn't have before. The second thing I noticed was these settings
were working fine right were they were, in the drawer like I had them.
So this is me trying to undo that and put things back to were they were.

@@ -473,9 +474,11 @@ org-journal
- changed journal type to a daily, mainly because there is no use in
loading all of it every start up. Like I have said before there seems
very little that I go back and look for when it comes to something like
a journal. There not a lot of top drawer material going into it so why
do I need it at my finger tips as soon as I can? There is very little of
use in the previous days journal. So there!

@@ -636,7 +639,14 @@ org-journal
- this is me finding Dooms settings and replacing my require setting.
was considering removing the whole lot but it complains right away about
my evil-surround-pairs-alist settings so not sure will have to try in
stages and see if there is a problem. Most times I have found that if
Doom has already set it. And you are ok with those settings you don't
need to put anything in your own config and just let doom do it. But in
this situation I may need something here. I also like to have the option
of changing something or doing a quick change to a package if need be or
if you need to add some variable you have a place to do that. So may
just leave here and experiment on different settings ... or not. will
see. For now it has been officially Doom-a-fide .

@@ -650,7 +660,7 @@ evil-surround
@@ -666,8 +676,10 @@ embrace
@@ -687,15 +699,35 @@ embrace
- there is a lot to unpack as to why this is here and what I am doing
with it.
- the short version is this is work in progess.
- The long version would have me going on a 25 line rant about
navigating code and parentheses. So lets just leave it at that.

@@ -1608,6 +1640,9 @@ added-pkgs
- this me first discovering "zone-packages" they amused me. So I thought
why not have them happen after a set amount of time idle? What could go
wrong he said with a grin. And ya there is a reason it is just an
interactive function. lesson learned. lets move on shall we.

@@ -1647,6 +1682,12 @@ my-keybindings
- added keybinding to open Doom-calendar. Cause why not?

@@ -1693,6 +1734,7 @@ my-keybindings
- two things going on here and one is that I learned that I can make a
function key a prefix key with out doing anything weird or added.
- the other has to do with still yet another video note taken app. I had
a hard on for something that would do this effortlessly. have yet to get
this worked out to a set manner. been monkeying around with all sort of
settings. But there is new one that fly under my radar. So needed to
make keybinding room for it. So a function-key prefix was my answer.

@@ -3157,3 +3198,17 @@ yequake
- this I like it replaced my conky I had set up and it is much easier to
manage.
- There is a problem when it comes to quit or kill it. It is suppose to
toggle but there has been times that it would not toogle. This caused a
problem in tring to just get rid of it after I have used it. The thing
that is making it harder to decide what is going on. Is the fact that
there is also a os (linux) keybinding that was set at the same time. a
bonus feature of it. But then when it ran a foul was I just pressing the
wrong keybinding? This is not hard to see someone doing this since it
has two to open it. one for emacs and one for linux. So when I had
problems could have this been the problem and the answer is .Ya it could
have been so now I just left every thing and see how it behaves the next
time and make sure I am hitting the right keys.