pin corfu and small but important fixes
@@ -1,7 +1,7 @@
- change this back to author like it was before

@@ -47,7 +47,7 @@ Package management
- a Corfu commit that broke evil-collections made me look at things a
little closer and by the way all the other 'load-path' the last
forward-slash should not be there.

@@ -199,7 +199,6 @@ dashboard
@@ -392,6 +391,7 @@ org-appearance
@@ -443,7 +443,6 @@ org-appearance
@@ -646,7 +645,6 @@ evil-surround
@@ -699,16 +697,6 @@ embrace
@@ -719,6 +707,9 @@ evil snipe
@@ -933,7 +924,7 @@ marginalia
- All these are clean up purge reorder

@@ -962,13 +953,10 @@ corfu
- this one is interesting in the fact that when I tried to grab the
whole function, when trying to mark it. the mark would stop at the
comment were there is a 'bracket' used in the comment. Did not like that
and thought that it could cause problems so the logical thing to do ....
That right. Purge. no real need for that comment.

@@ -977,6 +965,8 @@ corfu
- this just something I noticed that was missing from my set up so added
it just in case see if it did any thing. But was wanting to debug
earlier problem and thought it was worth a try and now I just left it.

@@ -987,12 +977,13 @@ corfu
- Here again in the proccess of debugging the problem I gave this
setting what I thought at the time as an improvement to make sure
history is working in Corfu
@@ the rest is clean up
* the bigger and more important changes are in the other files were I
pin corfu, get rid of compat and use doom's instead, find and point to
my proper personal dictionary. and many more