enough changes and additions warrants a update
- added a package as well as some fixes and personalised settings

@@ -136,6 +136,10 @@ startup-settings
- read about this setting on the web and it made sense to me to purge
this particular keybinding as to not hit unintentionaly.

@@ -407,8 +411,8 @@ org-capture
- This was a shock to me that basically this was not working the whole
time, but any way finally noticed that it was not sending anything I
captured to the file I intended. Insead it made another file with the
same name but no type of file. this was confusing as to why It would not
work. Turned out to be a typo and I had left a empty space between the
end of the word and the ending quote and this completely derailed
everything and did not notice till now. So fixed it finally

@@ -912,7 +915,8 @@ corfu
- replaced this setting for the second time. I want to use it but seem
to have to debug it seems shortly after this and it got removed for
debugging reasons. So I am going to try this again and see if it stays

@@ -1327,15 +1331,19 @@ flyspell-correct
- took a closer look at this setting and concluded that Doom is already
setting this and all I need is the keybinding but a closer look after
removing was that there was something using that same key when this was
not being used so I opted to use a remap in stead of a define-key
function. It seems fitting and worked thats the important part.

@@ -1557,7 +1565,26 @@ personal-functions
- This something that I had come across a similar function and I changed
to the way I thought it should be and it works like a charm I even can
use the defualt org-template to start it and don't have to worry what is
in it simply because it is a temporary file or buffer and does not
become a file until I save it. So this worked out good and enjoy it.

@@ -1942,6 +1969,18 @@ dired
- Here is another gem I came across. dired-subtree works just like one
would expect and don't know how it took me this long to find it or
notice it had actually seen it before but don't know why I did not use
it, seems like something I have been wanting for a while and it was
there the whole time and just skipped of it I guess, have it now thou.