many changes, adds, purges, reformatting, cleanup
- made enough changes the diff is getting hard to understand and needs
to be pushed and explained, these diffs don't really explain what is
going on or what was left after all the changes. I now see that some
need something better. I am looking at my own diffs and I am confused
and I was the one that did it. So ya I get it now why these diffs can be
more confusing then they really are.
- things on a whole seem to be working what I would say is better than
before and that has to do with some of the things I found and fixed or
did in a more agreeable way. So here is just the .config.org file
breakdown.

@@ -312,12 +312,11 @@
- this my attempt to get control of my org-templates. This has been
going on for awhile now and have only made small improvements.
- My thinking here is that something is tramping on this setting after I
have set it so I'm going to try an "after!" macro see if that improves
things or not as of right now I have to go to customise-menu, then
evaluate my changed block then when the changes show up write and save
them but even sometimes I can't get the menu to up date.

@@ -327,11 +326,9 @@
@@ -358,7 +355,7 @@
- these change come from the fact the previous changes broke it and did
not notice till I tried to use it again, the previous changes split the
command in the wrong spots and made it throw an error instead so this is
the fix to that and what made me look at this a little closer. The
simple fact that this was still a pain in the ass to work with made me
feel something needs to be changed here so it is easier to work with.
- It is working properly now, but may need to revisit it again in the
future to see if it improved with the changes.

@@ -430,7 +427,7 @@
- this was set wrong or copied wrong in any case it was wrong and it has
since been change back to what it was suppose to read and my journal was
changed to a daily a while back and just noticed this wrong variable.

@@ -740,14 +737,16 @@
- this is another long fucking story. So it starts with the fact that
flyspell has some problems when it comes to personal-dictionary or word
lists. This whole dictionary, language, environment, shit is convaluted
mess of bullshit, walls of text same sounding names then there is the
fact that we are dealing with every language and not only that but three
main dictionary type programs that are Ispell, aspell, then hundspell.
So everytime you have to mess with this you are greeted with walls of
text that go in to great detail about all things to do with it and it is
a mountain of shit to sift through just to find the little part that
pertains to you.
- So after that rant this is what is happening and it sounds like this
has been a problem on and off for a while and what happens is if you
save a word , the thinking is it won't fag that word next time but it
does immediately and then it tells you that it is spelled correct and it
won't move past it and won't stop flagging it as an error. Here the
thing I would go and simply try to remove this word from the dictionary
and would tell me no it can't it does not have the ability to do that.
??? So then I think alright I delete it my self. Ya good luck in trying
to find out were these words are kept there seems to be no file I can
find but it keeps on using it so I thought I need to get it to use my
own personal-dictionary that would be empty. Wholly old fuck did that
not happen right away fucked around for days yes days to get this
fucking thing to use the file I was pointing at in three different
places all the places I could find. But still refused. Final I got it to
use it after finding a setting in my .bashrc file that was cause some of
the head ache. Yet still it may use my dictionary now but still uses its
own after all that I now have a personal-dictionary but there is another
one it uses as well as mine to tell me certain words are misspelled but
when I try to correct them it tells me nothing to do the word is spelled
correct and that is it. It just stops there and keeps this file it is
using for this hidden from me. fucking aggravating.

@@ -763,18 +762,18 @@
- this is to revert a change I made a while back. it was wrong and I
miss understood what the 'm' ment in this. My bad

@@ -816,15 +815,6 @@
- purging this function because I do not use it and don't believe I
will.

@@ -857,6 +847,19 @@
- added yet anouther way to open a empty buffer I now have like 6 of
these all slightly different. Thinking here is some will be get used the
the other that don't get used can be removed after it has been
established which one gets used.

@@ -882,11 +885,6 @@
this was a package that sounded really useful. but then did not use it
hardly at all. So purge it and keep a record of it in the archive if
ever needed in the future.

@@ -900,22 +898,31 @@
@@ -959,7 +966,9 @@
@@ -968,6 +977,7 @@
- redid a couple of keybindings and added some others that I tend to use
added one that should help in the mini buffer to recall file names of
files I am working on, new way to open dictionary since I tend to forget
the ones I had so added some other to make it easier. the spilting of
code blocks I use but it seems to long and obscure that I made a easier
one.

@@ -1040,6 +1050,19 @@
- this function I got from Napoleon Wilson's Youtube-feed.
It and a setting in my operating system will allow me to grab youtube
links in a capture buffer that can be called independent of emacs.

@@ -1643,15 +1666,19 @@
- this was a reformatting, clean up the code.
not much added other than made it a macro '!' and enclosed the commands
in brackets. In some places it is not and only done if there are more
than one. So not sure if it was needed but will use it more see.
basically put as much as I could in the use-package function. He for
some reason put the keyword ':commands' in a weird place not were most
have it but at the bottom of the function. So I redid it and made sure
it still worked so. better.

@@ -1661,11 +1688,13 @@
- another reformatting

@@ -1709,18 +1738,19 @@
- this is another clean and reformatting of a section here it is denote
took an example of Jeremy cant remember last name in an article that was
about making a function to rename file in a denote fashion. Kept the
function as well in my notes to look at after when I may want this.

@@ -1762,16 +1792,15 @@
- Here is a good example of what I was talking about all I did here was
anouther reformatting after I noticed that there was a bracket missing
in one of the lines. But to look at this diff and figure out what was
removed and what if any thing was added I can't even tell. most of it
was simply moved to a different spot but it is hard to tell what is
going on here. even thou I was the one that did it.

@@ -1818,3 +1847,18 @@
- just now I realized that I'am using the wrong file this is the
config.el file not the org file thought those numbers were not maching
what I had in my head no I see this is the wrong file but fuck it not
doing this all over again for no good reason.
- This is me adding back me ediff settings after using it after I
realized I did like some of what it was doing so brought it back.