important changes purged line numbers, older settings
- *_note to self_ explain WHY! dam it!
The thing I learned was that when writing out this git-commit is to
explain why you made the change. Not that you changed something but why!
- I understand now how that is important in these commits, had it happen
to me I changed from using doom's mod config. to installing it
separately. Now I was thinking of changing it back, and then I
remembered that I already had it that way and changed it. Could not
remember why I had done that. So I looked back in the git and managed
after some time to track down were I made the change. And you know what
it said "here I changed this to that" and no reason as to why I had done
it. So from now on I will try to remember to explain why I did what I
did.
OK .... any who, these are the latest in changes and seem to have purged
some older stuff and ways or a least try to change them this is here so
if I need some of this back I can. So with that here are the changes as
from the config.org file

@@ -97,7 +97,8 @@ startup-settings
- This a big one for me, and it is directly related to some of the other
changes made further down, purged "line-numbers"
- still working on this as you will see further down, I have not yet
decided which prefix to use "spc" or "C-c a" the reason is doom uses the
space as leader key and has some of them blocked. I also made
keybindings that conflicted with it as well. most have been remedied.
But some will never be. example: is the use of "b" this is used for
buffer commands and I perfer that then wasting all those possible key
bindings just so I can use it for jumping around. But on the other hand
there may not be a need for certain ones. Really just looking for the
line jumping and the rest I might not need. I all ready use most of
these with out easy-motion, learned these back in the vim days thats a
long time ago. But my point is they still work with out the prefix to
bring up the avy menu. there are work arounds that work well enough.
and that is simply you use the same command and if there is more than
one you just hit "C-;" and the hint menu pops up allowing to jump to the
one you want. So that is why I will try out the different ways and try
to decide what works better may be I only need the line numbers and in
that case I could use the space since the "spc j/k" are open to be used
the other might not be that useful. Will try and work through this and
find out.

@@ -278,21 +279,6 @@ org-settings
- this was purged. and why it was is it was using up the "spc e" key and
I might need it for easy-motion, but this just got me to revisit this
and finally make a decision with this function which is a long story as
to why it is here and what I was doing with it, try to make it quick but
no promises. Starts out me come across a git hub page and it was for
changing any doc I was looking at into an org-file. This sounded like
something I would want and use. so installed and setup. Used it couple
times was happy with it. then realize some time later there was a built
in function which sounded like the same thing but could not tell if that
was the case or not. So planed on doing some trial and error thing to
see if they were doing the same or not. did not finish this tried just
could not make a decision on if there was a difference or not. So I just
left it. Now it was in the way and needed to be decided. and I said
purge it then if it seems like something I need I can bring it back. But
why have a package if there is a built in method?

@@ -599,7 +585,7 @@ evil snipe
- this is a more accurate header name.

@@ -610,9 +596,9 @@ evil snipe
- This were I will change the prefix like explained before

@@ -1403,33 +1389,6 @@ fly-spell
- This is an old function. I have since found the built in function that
will do the same so purge this to make it less confusing. Will use the
default method.

@@ -1583,10 +1542,6 @@ my-keybindings
- This is another one that gets removed because it is using a valuable
key ... the "k" is going to be needed to move up lines in easy-motion
- always wondered why certain keys were not being used for key binding
now I think I know why so they can be free to use them for easy-motion.

@@ -1628,9 +1583,10 @@ my-keybindings
- another long story in coming.
- tried to use "evil-embrace" found it clunky.
- the reason was two fold I had yet to figure out how to use
evil-surround to surround things with out selecting it first. I knew it
could be done but would forget when it came time to use it.
- the second was I thought it would help with code and parentheses. But
I found it confusing and clunky. So I looked harder into evil-surround
and seen what I was doing wrong so with that decided to purge this and
try the things I was going to use it for another way more built in
method.
- Same goes for the next one "sp-forward-slurp-sexp" this is something I
wanted to use but doom has this modified to not use the motion parts and
it has been neutered and most of it conficts with a lot of other stuff
and you will see were they turn it off for what ever reason in different
modes cause it interfere with some of it. as well it used some keys that
my linux distro uses so they were no good ether. The conclusion was
there was probably a better way that was already there I just needed to
figure it out. So purge this key and the setting in the init the loads
it.
- The next one is "transpose-char" another long story here. I just
learned to touch-type in the last couple years. The problem is this way
of teaching you self to type can lead to bad habits one is hair trigger
on the space bar. This is a problem, were as I found myself typing out
something hit the space then decide oh that is were I should end that
sentence. Then I would have to backup one space put the punctuation in
then hit the space again to carry on found myself doing this a lot.
became bother some. So I thought emacs probably can fix this for me.
with in a weeks time I had 3 to 4 maybe even 5 different ways to
accomplish this. This was one, hitting "transpose-char" will put the
punctuation back to the end of the word with out having to jump back
there. pretty slick, the problem is I all ready added a key binding
which is actually a keychord but it works well and got use to using it
in this situation it does do this it just jumps backward one space with
"j-n" this I don't move from the home row were as this key is two part
and not as easy to implement so will keep it around for awhile and see.

@@ -1864,35 +1820,32 @@ Dired
- here I am removing the dired line-numbers that I put there so then I
noticed that this section could use with some reformatting and cleaned
it up a bit.

@@ -3069,20 +3022,40 @@ yequake
- never finished this section out properly and thought most of this
could be consolidated into a use-package function and clean it up while
I was doing that I added another one and fix up the files it showed.

@@ -3095,6 +3068,8 @@ yequake
added some links to the readme for next time they are handy to have
right here in the config so going to keep doing this.

@@ -3109,6 +3084,22 @@ ediff
- added a keybinding chart to help remind me of some of them.