most of this is just reordering
@@ -166,6 +166,11 @@ dashboard
- moved to the left since use this more than the others.

@@ -597,8 +597,8 @@ evil snipe
- here I am experimenting with using 'spc' as prefix key for
evil-easymotion. The thinking here is I mainly need it for lines and can
use the other motions with out a prefix. So 'spc' is easier to hit than
'C-c a'. But to be honest there is no reason I can't use both and use
the 'C-c a' when I can't use 'spc'. That all well and good but I need
answers.

@@ -612,30 +612,6 @@ key chords
- re-organized some header to make small groups this was one I simply
move by the other keybindings.

@@ -3041,22 +3041,6 @@ yequake
- reformatted this section by putting it all in a use-package function