added pkg,function rewrite success,cleanup
- added hnreader to read hacker news, it is more logical to review a
thread before opening it. So that was the reason, but there is more
going on here then just that.
- the more that I speak of is the story of me putting my first function
up on the "Emacs Hoist" add got my hands dirty. It is proably an old
story, but it is my first and would like to pump my own tires this one
time.
- Will try to summarize it the best I can.
 - read this blog.
 =>https://blog.dornea.nu/2023/04/21/read-hackernews-and-reddit-the-emacs-way/
 - used the riddit function and worked well enough that I thought I
 would want to do the hacker function, my first attempt. failed. The
 reason had to do with my limited knowledge of emacs in general. So I
 just used the one that did work and remove and archive the other.
 - After many months go by and I start to pick up on a few things and
 gather some knowledge. I think lets try that again this time I think I
 know what I was doing wrong, I was right, but the thing was it still
 would not work. This ended after several frustrating tries. So once
 again had to undo everthing and archive what I had. just leave it be.
 - Fast forward to now and I can say I know more now then I did then so
 it was time to try it yet again for the third time. Very similar thing
 happens at the first install. just keep getting an error. But this time
 I am better a deducing the problem and by doing some quick eval of the
 functions being used I could tell that they worked on their own so it
 had to be the data that was be used was the problem. This is were I
 stared using that knowledge of using the help section to be able to
 figure out what was going wrong. Was even smart enough to replace the
 last function to "print" command and this told me what it was
 collecting and how this was the problem. It was no longer grabbing
 anything it could not find anything so I modified it and was able to
 get it to grab the link. The problem was in was in the form of a con
 cell, and would not work by itself. This is were I had to go a learn
 the basics of con cells and right away it was apparent that I should be
 able to grab just the link so back to the drawing board and try to come
 up with a version that only grab the cdr of the con cell. first attempt
 did not work cause I just tried to add the command to the end of the
 function which did not work so after staring at it I realized what I
 needed to do and it worked. I could see by my print function that it
 was now just grabbing the url at the end of the con cell and it worked
 after reassembling the function with my new modification. I gave it a
 try and worked like a charm. So very happy about this. Feels good to be
 able to have the knowledge to do it and have it work feels like I
 leveled up my Emacs-Foo! So good on me. Yaaa!
 - the list of things I learned in doing this would be long.
 - here is the lastest changes and my shiny new function that I put
 together myself. left the old commented not working function for
 comparison.
@@ -1465,6 +1465,9 @@ added-pkgs
- This is the pkg that I was talking about, the readme said nothing
about declaring it in your config and think the functions will work when
called from m-x but do not work for elfeed on less you use a "require
pkg" this will add some load time so I went for the "use-package"
command instead. had to check to make sure it would still work some do
not need to be installed fully to work but was able to get this to work
just could not use a ":defer t" option so opted for an ":after:"
instead.
- This allows the elfeed-function to work

@@ -1984,9 +1987,9 @@ web-url-handlers
- This had extra cruf on it that it did not need so "purge"

@@ -1994,14 +1997,20 @@ web-url-handlers
- This chunk has a lot going on even thou it don't look like it.
- first I will tell "The Why" then fill in the minor story behide it.
- "The why": use a similar setting for 'all' video web sites and think
it is better than having a function and keybinding. were as this works
just by hitting enter. So have enjoyed this function and I thought it
could be used here as just as well. I was right. worked before I got the
elfeed-search function to work. but that is a long story
- "The story" Warning the opinions expressed are solely my own and not
based on any fact and could be a simple lack of communicating and me not
understanding how to set "custom-variables" in my config. read how it is
suggested to do this and was wanting to do it as well.But FUCK ME.
... really! I must of read the first paragraph of the custom-file. Like
20 times and still not get it. What I mean is I understood it to mean
don't edit that file cause it may cause an error and it producing
another default file and this would brick your system until it was
removed but if you could not start it you may never know that was the
problem, and that was what the warning at the top meant. But what I
found out the hard way was. NO THAT'S NOT IT. It was you could not use
that particular cammand any were else. Me trying to get this variable in
my config to set it self. eventually evaluated it in the scratch-buffer
only to notice my whole custom-file was now replaced with only that one
setting and no way I could see of getting this back. Not a happy camper
at this moment. But luckily for me I had did a git push only 4 days
prior and used vc-revert and was able to get things back. Did not care
for the experience and would not recommend it. So after that it was back
to being happy to just eval, then open the costume-menu and save it.

@@ -2075,11 +2084,11 @@ deft
- Here decided I wanted access to the elfeed use-package function but
was just letting doom do it. So copied what doom was using since that
seemed like the optimal setting and was all ready using so seemed
logical just put it in my config. so that is what happened here had it
listed under a no tangle header then just moved it to my personal
setting and used it as my own slightly modified.

@@ -2087,16 +2096,11 @@ deft
- did not think this long comment was needed. self explanatory.

@@ -2146,34 +2150,6 @@ elfeed-org
- This was the same as elfeed want access to the use-package function so
copied what Doom was already using. But put it in the config.

@@ -2255,6 +2231,30 @@ Elfeed *This is the one.* Just take look at it.
just beaming like a happy, shiny, new-function. All happy, shiny and
shit.
- So the why is, that it is a better interface than the browser and i
like it better.
- the what is, the top one is the one from the blog used that did not work
for me. The one under it is the one I produced. Works as well as can be
expected. It is better because I crafted it my self. So wanted to
document it.

@@ -2281,14 +2282,18 @@ Elfeed
- also added the elfeed-goodies use-package function

that all for now stay tuned for the next installment of "The config files"